Temporary check to make users cannot spoof into other user's data
This commit make changes to the JWT issuing process as to not leak too much sensitive information. Also, this change tosses in a check as a stopgap measure to make sure users cannot spoof their way into other user's data. A better solution is preferable as right now we just making sure the path parameter and the jwt claim are the same any nothing more.