Added public card route
This commit adds a route to allow any public facing client to access VCard/Profile data provided they have a valid profileId/cardId. Note that id refers to the generated shortcodes and not the ObjectIDs generated by Mongod. In addition, this makes a few changes in other routes that manually set and sent error responses to instead use the error handlers now defined in the app.js