Refactored PATCH to workaround remove op with Mongoose Models
This commit changes how JSON Patches are applied to documents; the change largely takes remove operations and changes them into replace operations that set the values as undefined. Prior to this, JSON Patches were applied using the applyPatch method. However, while the patch would succeed, remove operations would not actually remove the key/field from the document. Upon research, the correct way to do this is to set field in question as 'undefined'. However, 'undefined' is not valid in JSON and therefore could not be generated in the future. Therefore, this switch is done server side instead

Additionally, some refactoring was done to the conditionals to make them more streamlined and easier to read by using the guard clase pattern