Changes response body to safer data
This commit changes the response object for the POST route on /users/new to simply be just include a meta object and success object containing creation data and username. This change was done as previously the direct serialization of the user Mongoose document may have revealed too much to potential attackers.