Added Secure flag to JWT cookie
This commit largely just adds the secure option for JWT cookie storage> Work still needs to be done to see CSRF cookie as httpOnly and Secure as well but largely this ready. With local development ready, we are largely ready to merge. Note that we will be using custom headers to send the CSRF tokens as it is cleaner and given that we control the api we make that happen; custom header will be X-CSRF-TOKEN